uploads/requirements/requirements_dev.txt
# Dependencies with versioning, this was generated by the
# command "pip freeze > requirements.txt"

##############################################################################
Django==1.11.5              # Backend web framework
djangorestframework==3.6.4  # provides rest_framework django app
psycopg2==2.7.3.1           # Library to talk to postgresql
pytz==2017.2                # automatic dependency, for global timezones
autoflake==0.7		        # library to suppress unused imports
flake8==3.4.1		        # syntax checking package
jedi==0.11.0		        # dependency for anaconda-mode (emacs/spacemacs)
json-rpc==1.10.3	        # dependency for anaconda-mode (emacs/spacemacs)
mccabe==0.6.1
nose==1.3.7		            # Testing framework
parso==0.1.0
pycodestyle==2.3.1
pyflakes==1.5.0
service-factory==0.1.5      # Required for Django RESTful API
django-cors-headers==2.1.0  # Required for making cross domain requests
attrs==17.2.0
autoflake==0.7
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
coverage==4.4.1
enum34==1.1.6
httpie==0.9.9               # test RESTful API from command line
hypothesis==3.32.1          # quick-check like testing
idna==2.6
jedi==0.11.0
json-rpc==1.10.3
Pygments==2.2.0             # syntax highlighting
requests==2.18.4
urllib3==1.22
django-extensions==1.9.1    # provides useful utils like ./manage.py show_urls
django-rest-auth==0.9.2	    # Required for rest_auth urls, and authentication.
django-allauth==0.33.0      # Create user and login urls.
oauthlib==2.0.5             # User auth
python-openid==2.2.5
requests-oauthlib==0.8.0
six==1.11.0
py4j=0.10.6
##############################################################################
